feat(plasmic): split state from context
